refactor: move one more repl class
